requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.3
Django==4.2.2
django-request==1.6.2
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
Werkzeug==2.3.6
